Update GitHub and docs links to specific pages
At the moment, it's difficult to figure out how to use the tool if you land on it directly.

This PR updates the GitHub and docs links to the specific GitHub repo and specific blog post announcing it (rather than your general GitHub profile/blog main page). This should help new users discover the documentation and codebase more easily.